Refactor user module to be firebase module
