+++ src/store/index.js
@@ -3,7 +3,7 @@ import Vuex from 'vuex'
 // import * as actions from './actions'
 // import * as getters from './getters'
 import app from './modules/app'
-import user from './modules/user'
+import firebase from './modules/firebase'
 import sound from './modules/sound'
 
 Vue.use(Vuex)
@@ -15,7 +15,7 @@ export default new Vuex.Store({
   // getters,
   modules: {
     app,
-    user,
+    firebase,
     sound
   },
   strict: debug
